Remove redundant v0 APIs
