+++ src/applications/vaos/tests/appointment-list/components/AppointmentsPage.unit.spec.js
@@ -15,8 +15,6 @@ import {
   getTestDate,
 } from '../../mocks/setup';
 import AppointmentsPage from '../../../appointment-list/components/AppointmentsPage';
-import { mockPastAppointmentInfo } from '../../mocks/helpers';
-import { createMockAppointmentByVersion } from '../../mocks/data';
 import { mockVAOSAppointmentsFetch } from '../../mocks/helpers.v2';
 import { getVAOSRequestMock } from '../../mocks/v2';
 
@@ -98,8 +96,6 @@ describe('VAOS Page: AppointmentsPage', () => {
 
   it('should display updated upcoming appointments page', async () => {
     // Given the veteran lands on the VAOS homepage
-    mockPastAppointmentInfo({});
-
     // When the page displays
     const screen = renderWithStoreAndRouter(<AppointmentsPage />, {
       initialState,
@@ -247,21 +243,6 @@ describe('VAOS Page: AppointmentsPage', () => {
 
   it('should display updated past appointment page', async () => {
     // Given the veteran lands on the VAOS homepage
-    const pastDate = moment().subtract(3, 'months');
-    const data = {
-      id: '1234',
-      kind: 'clinic',
-      clinic: 'fake',
-      start: pastDate.format(),
-      locationId: '983GC',
-      status: 'booked',
-    };
-    const appointment = createMockAppointmentByVersion({
-      version: 0,
-      ...data,
-    });
-    mockPastAppointmentInfo({ va: [appointment] });
-
     // When the page displays
     const screen = renderWithStoreAndRouter(<AppointmentsPage />, {
       initialState,
@@ -329,8 +310,6 @@ describe('VAOS Page: AppointmentsPage', () => {
     };
 
     // Given the veteran lands on the VAOS homepage
-    mockPastAppointmentInfo({});
-
     // When the page displays
     const screen = renderWithStoreAndRouter(<AppointmentsPage />, {
       initialState: defaultState,
@@ -354,8 +333,6 @@ describe('VAOS Page: AppointmentsPage', () => {
 
     it('should display updated title on upcoming appointments page', async () => {
       // Given the veteran lands on the VAOS homepage
-      mockPastAppointmentInfo({});
-
       // When the page displays
       const screen = renderWithStoreAndRouter(<AppointmentsPage />, {
         initialState: defaultState,
@@ -476,21 +453,6 @@ describe('VAOS Page: AppointmentsPage', () => {
 
     it('should display updated past appointments page title', async () => {
       // Given the veteran lands on the VAOS homepage
-      const pastDate = moment().subtract(3, 'months');
-      const data = {
-        id: '1234',
-        kind: 'clinic',
-        clinic: 'fake',
-        start: pastDate.format(),
-        locationId: '983GC',
-        status: 'booked',
-      };
-      const appointment = createMockAppointmentByVersion({
-        version: 0,
-        ...data,
-      });
-      mockPastAppointmentInfo({ va: [appointment] });
-
       // When the page displays
       const screen = renderWithStoreAndRouter(<AppointmentsPage />, {
         initialState: defaultState,
