+++ src/applications/vaos/tests/new-appointment/components/ReviewPage/index.cc-request.unit.spec.js
@@ -21,8 +21,6 @@ import {
 import ReviewPage from '../../../../new-appointment/components/ReviewPage';
 
 import { mockAppointmentSubmitV2 } from '../../../mocks/helpers.v2';
-import { createMockCheyenneFacilityByVersion } from '../../../mocks/data';
-import { mockFacilityFetchByVersion } from '../../../mocks/fetch';
 
 const initialStateVAOSService = {
   featureToggles: {
@@ -260,12 +258,6 @@ describe('VAOS Page: ReviewPage CC request with VAOS service', () => {
       },
     });
 
-    mockFacilityFetchByVersion({
-      facility: createMockCheyenneFacilityByVersion({
-        version: 0,
-      }),
-      version: 0,
-    });
     setFetchJSONFailure(
       global.fetch.withArgs(`${environment.API_URL}/vaos/v2/appointments`),
       {
