+++ src/applications/vaos/tests/new-appointment/components/TypeOfCarePage.unit.spec.js
@@ -11,7 +11,6 @@ import { mockFetch, setFetchJSONResponse } from 'platform/testing/unit/helpers';
 import moment from 'moment';
 import environment from '@department-of-veterans-affairs/platform-utilities/environment';
 import { createTestStore, renderWithStoreAndRouter } from '../../mocks/setup';
-import { mockParentSites } from '../../mocks/helpers';
 import {
   mockVAOSParentSites,
   mockV2CommunityCareEligibility,
@@ -100,7 +99,6 @@ describe('VAOS Page: TypeOfCarePage', () => {
 
   it('should show type of care page with all care types', async () => {
     const store = createTestStore(initialState);
-    mockParentSites(['983'], []);
     mockV2CommunityCareEligibility({
       parentSites: [],
       supportedSites: [],
@@ -155,7 +153,6 @@ describe('VAOS Page: TypeOfCarePage', () => {
 
   it('should save type of care choice on page change', async () => {
     const store = createTestStore(initialState);
-    mockParentSites(['983'], []);
     mockV2CommunityCareEligibility({
       parentSites: [],
       supportedSites: [],
@@ -202,7 +199,6 @@ describe('VAOS Page: TypeOfCarePage', () => {
   });
   it('should not allow users who are not CC eligible to use Podiatry', async () => {
     const store = createTestStore(initialState);
-    mockParentSites(['983'], []);
     mockV2CommunityCareEligibility({
       parentSites: [],
       supportedSites: [],
